Phpstan level 8 OK

+++ Src/Db/DbDeleteService.php
@@ -13,6 +13,10 @@ class DbDeleteService extends DbConnexion {
     public function deleteTrajet(int $id) { 
         try {
         $connexion = $this->connexion(1);
+        if (!$connexion instanceof \PDO) {
+                
+                throw new \Exception("La connexion à la base de données a échoué.");
+            }
         $sql = "delete from trajet where id = :id";
         $query = $connexion->prepare((string) $sql);
         $query->execute(['id' => $id]);
@@ -28,7 +32,12 @@ class DbDeleteService extends DbConnexion {
     */
     public function deleteVille(int $id) {
         try {
+            
             $connexion = $this->connexion(2);
+            if (!$connexion instanceof \PDO) {
+                
+                throw new \Exception("La connexion à la base de données a échoué.");
+            }
             $sql = "delete from ville where id = :id";
             $query = $connexion->prepare((string) $sql);
             $query->execute(['id' => $id]);
